move PlanResult to types
