+++ include/smac_planner/smac_planner_hybrid.hpp
@@ -61,75 +61,6 @@ public:
     std::string name,
     costmap_2d::Costmap2DROS* costmap_ros) override;
 
-    struct PlanResult {
-      uint32_t result_code = mbf_msgs::GetPathResult::SUCCESS;
-      double cost = 0;
-      std::string message = "";
-
-      // constructor
-      PlanResult() = default;
-      PlanResult(uint32_t result_code, double cost, const std::string& message)
-        : result_code(result_code), cost(cost), message(message) {}
-
-      // Getter for path
-      const std::vector<geometry_msgs::PoseStamped>& path() const {
-        return _path;
-      }
-
-      // Setter for path
-      void setPath(const std::vector<geometry_msgs::PoseStamped>& new_path) {
-        _path = new_path;
-        _length = Utils::length(_path);
-      }
-
-      // Getter for length
-      double length() const {
-        return _length;
-      }
-
-      bool isValid() const
-      {
-        return result_code == mbf_msgs::GetPathResult::SUCCESS;
-      }
-
-      PlanResult operator+(const PlanResult& other_result) const {
-        if (!isValid() || !other_result.isValid()) {
-          return PlanResult{
-            mbf_msgs::GetPathResult::FAILURE,
-            0.0,
-            "One of the segments is invalid, cannot add paths"
-          };
-        }
-
-        PlanResult combined = *this;
-        combined.cost += other_result.cost;
-
-        const auto& other_path = other_result.path();
-        std::vector<geometry_msgs::PoseStamped> new_combined_path = path();
-
-        if (!new_combined_path.empty() && !other_path.empty() &&
-            new_combined_path.back() == other_path.front()) {
-          new_combined_path.insert(
-            new_combined_path.end(),
-            other_path.begin() + 1,
-            other_path.end()
-          );
-        } else {
-          new_combined_path.insert(
-            new_combined_path.end(),
-            other_path.begin(),
-            other_path.end()
-          );
-        }
-
-        combined.setPath(std::move(new_combined_path));
-        return combined;
-      }
-
-      private:
-        double _length = 0;
-        std::vector<geometry_msgs::PoseStamped> _path{};
-    };
 
     /**
    * @brief Creating a path from start to goal pose based on params.
